Separation of files by algorithm type
