+++ include/xsimd_algorithm/algorithms.hpp
@@ -8,24 +8,11 @@
  *                                                                          *
  * The full license is in the file LICENSE, distributed with this software. *
  ****************************************************************************/
-#ifndef EMSCRIPTEN
-#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
-#include "doctest/doctest.h"
-#else
 
-#define DOCTEST_CONFIG_IMPLEMENT
-#include "doctest/doctest.h"
-#include <emscripten/bind.h>
+#ifndef XSIMD_ALGORITHMS_HPP
+#define XSIMD_ALGORITHMS_HPP
 
-int run_tests()
-{
-    doctest::Context context;
-    return context.run();
-}
-
-EMSCRIPTEN_BINDINGS(my_module)
-{
-    emscripten::function("run_tests", &run_tests);
-}
+#include "xsimd_algorithm/stl/reduce.hpp"
+#include "xsimd_algorithm/stl/transform.hpp"
 
 #endif
